my hrm: job details view
